Updated for OR2019 Hamburg
The order of the exercises has been revisited as many content endpoints are now restricted to administrator. A new exercise 5 to explore the submission and workflow capabilities has been introduced. The postman tips section has been enhanced with information about the use of environments and a default configuration included for download